read_view: use tuple_dictionary to store field names
We don't need to know field types and constraints in read view. All we
need is field names. Currently, we use a field_def array to store field
names, but it's inconvenient to deal with it. To handle the case when a
read view is created while space upgrade is in progress, we'll need to
store old field names in addition to new field names so we'd have to use
two field_def arrays, which would be rather cumbersome. Let's switch to
tuple_dictionary for simplicity.

Actually, the only reason why we used a field_def array instead of
tuple_dictionary is the absence of tuple_dictionary_dup() function.
This commit implements this function. To avoid code duplication, we
refactor tuple_dictionary_new(). We also switch it to grp_alloc and
xmalloc for convenience and drop tuple_dictionary_delete_hash() because
it's redundant (it's okay to delete the hash right away without deleting
all entries from it first).

NO_DOC=refactoring
NO_TEST=refactoring
NO_CHANGELOG=refactoring